Support JSON payload on post to saveClient
Unable to POST using FormData and XHR, I needed to use JQuery .ajax().

Also, due to my insistence to comply with the JSON format across the Java, node.js, and Ruby versions of the application, I could not
use JQuery's .serializeArray() to serialize all form fields to JSON in a way that supports camel-case in the Clientele class. Resorted
to manually building the JSON for each form field.

Also discovered the ThymeLeaf supports HTML-5's datetime-local input type, as long as th:field is used and not th:value.